test: fix start-local-chain-boot
