+++ packages/inter-protocol/scripts/start-local-chain-boot.js
@@ -3,34 +3,14 @@
 import { buildRootObject as buildPsmRootObject } from '@agoric/vats/src/core/boot-psm.js';
 import { CHAIN_BOOTSTRAP_MANIFEST } from '@agoric/vats/src/core/manifest.js';
 import * as utils from '@agoric/vats/src/core/utils.js';
-import { E, Far } from '@endo/far';
+import { Far } from '@endo/far';
 import {
+  installPriceAggregatorContract,
   PRICE_FEEDS_MANIFEST,
   startPriceFeeds,
 } from '../src/proposals/price-feed-proposal.js';
 
 /** @typedef {import('@agoric/inter-protocol/src/proposals/econ-behaviors.js').EconomyBootstrapSpace} EconomyBootstrapSpace */
-
-/** @param {BootstrapSpace & { devices: { vatAdmin: any }, vatPowers: { D: DProxy }, }} powers */
-const installPriceAggregatorContract = async ({
-  vatPowers: { D },
-  devices: { vatAdmin },
-  consume: { zoe },
-  installation: {
-    produce: { priceAggregator },
-  },
-}) => {
-  // Copied from installBootContracts:
-
-  // This really wants to be E(vatAdminSvc).getBundleIDByName, but it's
-  // good enough to do D(vatAdmin).getBundleIDByName
-  const bundleCap = D(vatAdmin).getNamedBundleCap('priceAggregator');
-
-  const bundle = D(bundleCap).getBundle();
-  // TODO (#4374) this should be E(zoe).installBundleID(bundleID);
-  const installation = E(zoe).install(bundle);
-  priceAggregator.resolve(installation);
-};
 
 /**
  * Build root object of the PSM-only bootstrap vat.
@@ -76,8 +56,10 @@ export const buildRootObject = (vatPowers, vatParameters) => {
     console.log('awaiting startPriceFeeds');
     await Promise.all([
       psmRootObject.bootstrap(vats, devices),
-      installPriceAggregatorContract(powersFor('installBootContracts')),
-      startPriceFeeds(powersFor('startPriceFeeds'), {
+      installPriceAggregatorContract(
+        powersFor(installPriceAggregatorContract.name),
+      ),
+      startPriceFeeds(powersFor(startPriceFeeds.name), {
         options: { demoOracleAddresses },
       }),
     ]);
